initial files for the project
This include the files that allow connection and retrieval/set data to the redis in memory database. unlike other framework that I found online, this one do not implements the redis API in term of redis CLI, but more in STL like interface, which allow for the user to use REDIS just like its an STL container. The one exception is the channel concept that is used to register for events and sending them.